:truck: Move to utils
