+++ package.json
@@ -32,7 +32,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "build-test": "yarn build && serve -s build"
   },
   "eslintConfig": {
     "extends": "react-app"
